+++ ts/examples/vite-browser-render/index.html
