+++ ts/examples/vite-browser-render/package.json
@@ -11,7 +11,8 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@live-compositor/browser-render": "file:../../../@live-compositor/browser-render",
+    "@live-compositor/browser-render": "file:../../@live-compositor/browser-render",
+    "mp4box": "^0.5.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
