+++ ts/examples/vite-browser-render/src/examples/Counter.tsx
@@ -1,8 +1,8 @@
-import { useEffect, useRef, useState } from 'react';
-import './App.css';
-import { Renderer, loadWasmModule } from '@live-compositor/browser-render';
 
-function App() {
+import { useEffect, useRef, useState } from 'react';
+import { useRenderer } from './utils';
+
+function Counter() {
   const canvasRef = useRef<HTMLCanvasElement | null>(null);
   const [count, setCount] = useState(0);
   const renderer = useRenderer();
@@ -62,7 +62,6 @@ function App() {
 
   return (
     <>
-      <h1>Renderer Example</h1>
       <div className="card">
         <canvas ref={canvasRef} width={300} height={300}></canvas>
         <button onClick={() => setCount(count => count + 1)}>Count +1</button>
@@ -71,30 +70,4 @@ function App() {
   );
 }
 
-function useRenderer(): Renderer | null {
-  const [renderer, setRenderer] = useState<Renderer | null>(null);
-  useEffect(() => {
-    const setupRenderer = async () => {
-      await loadWasmModule('./assets/live-compositor.wasm');
-      const renderer = await Renderer.create({
-        streamFallbackTimeoutMs: 500,
-      });
-
-      await renderer.registerImage('img', {
-        asset_type: 'gif',
-        url: 'https://media.tenor.com/eFPFHSN4rJ8AAAAM/example.gif',
-      });
-      await renderer.registerFont(
-        'https://fonts.gstatic.com/s/notosans/v36/o-0mIpQlx3QUlC5A4PNB6Ryti20_6n1iPHjcz6L1SoM-jCpoiyD9A-9a6Vc.ttf'
-      );
-
-      setRenderer(renderer);
-    };
-
-    setupRenderer().catch(err => console.error(err));
-  }, []);
-
-  return renderer;
-}
-
-export default App;
+export default Counter;
