+++ ts/examples/vite-browser-render/src/examples/mp4/mp4box.d.ts
