+++ ts/examples/vite-browser-render/src/main.tsx
