+++ ts/examples/vite-browser-render/src/vite-env.d.ts
