+++ ts/examples/vite-browser-render/tsconfig.app.json
