+++ ts/examples/vite-browser-render/vite.config.ts
