--- ts/examples/web-examples/mp4/.eslintignore
@@ -1,1 +1,0 @@
-dist
