--- ts/examples/web-examples/mp4/.eslintrc.json
@@ -1,5 +1,0 @@
-{
-  "extends": [
-    "../../../.eslintrc.base.json"
-  ]
-}
