--- ts/examples/web-examples/mp4/index.html
@@ -1,27 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8" />
-  <link rel="stylesheet" href="./src/style.css">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>LiveCompositor MP4 example</title>
-</head>
-
-<body>
-  <div id="app">
-    <canvas id="canvas"></canvas>
-  </div>
-  <script type="module">
-    import {play} from './src/mp4';
-
-    async function main() {
-      const BUNNY_URL = 'https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4';
-      play(BUNNY_URL);
-    }
-
-    main();
-  </script>
-</body>
-
-</html>
