--- ts/examples/web-examples/mp4/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "web",
-  "version": "0.0.0",
-  "devDependencies": {
-    "typescript": "^5.5.3",
-    "vite": "^5.4.1",
-    "vite-plugin-static-copy": "^1.0.6"
-  },
-  "private": true,
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "lint": "eslint .",
-    "typecheck": "tsc --noEmit",
-    "preview": "vite preview"
-  },
-  "type": "module",
-  "dependencies": {
-    "mp4box": "^0.5.2"
-  }
-}
