--- ts/examples/web-examples/mp4/vite.config.ts
@@ -1,22 +1,0 @@
-import { defineConfig } from 'vite';
-import { viteStaticCopy } from 'vite-plugin-static-copy';
-import { createRequire } from 'node:module';
-
-const require = createRequire(import.meta.url);
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [
-    viteStaticCopy({
-      targets: [
-        {
-          src: require.resolve('@live-compositor/browser-render/dist/live-compositor.wasm'),
-          dest: 'assets',
-        },
-      ],
-    }),
-  ],
-  optimizeDeps: {
-    exclude: ['@rollup/browser'],
-  },
-});
